refactor(tui): consolidate centering, kv-row, and truncation helpers
Pure no-op consolidation (golden renders byte-identical):
- Delete the byte-identical centered_rect fork in hf_dialog; add one
  parameterized layout::centered_abs replacing the two divergent
  confirm/help `centred` helpers (each keeps its exact margin).
- Promote settings::kv/kv_focused to tui/fmt::kv_row/kv_row_focused.
- Centralize truncation in tui/fmt (truncate_end + truncate_start);
  they keep distinct semantics since end- vs left-truncation render
  differently — not force-merged.

Left panel_block and the hand-rolled pane blocks unchanged: none is
byte-identical to an extended panel_block, so migrating would alter
renders. That cross-pane block/padding unification is a visual-polish
follow-up, not a no-op consolidation.
